+++ MedicalAppointmentBookingSystem/src/main/java/org/mabs/repository/PrescriptionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface PrescriptionRepository extends JpaRepository<Prescription, Integer> {
-    List<Prescription> findByMedicalRecord_Id(Integer id);
+public interface PrescriptionRepository extends JpaRepository<Prescription, Long> {
+    List<Prescription> findByMedicalRecord_Id(Long id);
 }
